double kompatibilis

+++ HarmadikOra/Program.cs
@@ -8,23 +8,23 @@ namespace HarmadikOra
 {
     class Program
     {
-        static int[] Input(int darabszam)
+        static double[] Input(int darabszam)
         {
             string input = "";
             Console.Write("Adjon meg az előbb megadott darbmennyiségű számot (space-ekkel elválasztva): ");
             input = Console.ReadLine();
             string[] inputSplit = input.Split(' ');
-            int[] tomb = new int[darabszam];
+            double[] tomb = new double[darabszam];
             for (int i = 0; i < tomb.Length; i++)
             {
-                tomb[i] = Convert.ToInt32(inputSplit[i]);
+                tomb[i] = Convert.ToDouble(inputSplit[i]);
             }
             return tomb;
         }
 
-        static int Max(int[] tomb)
+        static double Max(double[] tomb)
         {
-            int max = tomb[0];
+            double max = tomb[0];
             for (int i = 1; i < tomb.Length; i++)
             {
                 if (tomb[i] > max)
@@ -35,9 +35,9 @@ namespace HarmadikOra
             return max;
         }
 
-        static int Min(int[] tomb)
+        static double Min(double[] tomb)
         {
-            int min = tomb[0];
+            double min = tomb[0];
 
             for (int i = 1; i < tomb.Length; i++)
             {
@@ -56,7 +56,7 @@ namespace HarmadikOra
             {
                 Console.Write("Adja meg az elemszámot: ");
                 int darabszam = Convert.ToInt32(Console.ReadLine());
-                int[] tomb = Input(darabszam);
+                double[] tomb = Input(darabszam);
                 Console.WriteLine("A legnagyobb elem: " + Max(tomb));
                 Console.WriteLine("A legkisebb elem: " + Min(tomb));
                 Console.ReadKey();
